pyproject.toml
[project]
name = "saas-researcher"
version = "0.1.0"
description = "This project is describe how to use the different AI Agents framework and tools, their trade off, when to use what, best practices, how to deploy your agents and monitor them in production"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "pillow>=12.1.0",
    "matplotlib>=3.8.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "duckduckgo-search>=6.0.0",
    "python-dotenv>=1.0.0",
    "ipython>=9.8.0",
    "ddgs>=9.10.0",
]
